requirements.txt
Django==3.2.7
flake8==3.9.2
pokebase==1.3.0
djangorestframework==3.12.4
